src/backend/llm_service/requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
fastapi==0.115.4
h11==0.14.0
ibm-cloud-sdk-core==3.22.0
ibm-watson==8.1.0
idna==3.10
pydantic==2.9.2
pydantic_core==2.23.4
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.32.3
six==1.16.0
sniffio==1.3.1
starlette==0.41.2
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.32.0
websocket-client==1.8.0
pygame
google
google-generativeai
python-magic
python-magic-bin
python-multipart